requirements.txt
langgraph>=0.4.0
akshare>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.27.0
duckduckgo-search>=4.0.0
pydantic>=2.0.0
pytz>=2024.1
apscheduler>=3.10.0
